User and Validator class Update and integration - removed the user_type attribute from the user class (to be discussed later) - renamed abstractuser.py to user.py - made slight adjustment to validator.py - integrated validate email and validate input into user class setters
