+++ src/scms/validator.py
@@ -1,13 +1,13 @@
 import re
 
 class Validator:
-    def __init__(self):
-        self.email_pattern = re.compile(
+    @staticmethod
+    def validate_email(email: str) -> bool:
+        email_pattern = re.compile(
             r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
         )
+        return bool(email_pattern.match(email))
 
-    def validate_email(self, email: str) -> bool:
-        return bool(self.email_pattern.match(email))
-
-    def validate_input(self, input_str: str) -> bool:
+    @staticmethod
+    def validate_input(input_str: str) -> bool:
         return bool(input_str) and input_str.strip() != ""
